using long instead of decimal for Value and Fee

+++ source/.net/HackChain/HackChain.Core/Model/Account.cs
@@ -3,7 +3,7 @@
     public class Account
     {
         public string Address { get; set; }
-        public decimal Balance { get; set; }
+        public long Balance { get; set; }
         public long Nonce { get; set; }
     }
 }
